Migrate and disable tunneling plugins
