+++ PVGeo/__init__.py
@@ -7,7 +7,7 @@ from . import grids
 from . import gslib
 from . import model_build
 from . import readers_general
-from . import tunneling
+#TODO: from . import tunneling
 from . import ubc
 
 from . import _helpers
