--- src/filters-general/build.config
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<!-- This is the plugin group configuration for the general filters group -->
-<!-- name: The name/label used for the group in ParaView's plugin manager -->
-<!-- category: The filter/source menu category in ParaView. This attribut will override what is set as FilterCategory in the plugin .py file.-->
-<PluginSuite
-    name="PVGeo General Filters"
-    category="PVGeo General Filters">
-</PluginSuite>
