--- src/readers-general/build.config
@@ -1,7 +1,0 @@
-<!-- This is the plugin group configuration for the general filters group -->
-<!-- name: The name/label used for the group in ParaView's plugin manager -->
-<!-- category: The filter/source menu category in ParaView. This attribut will override what is set as FilterCategory in the plugin .py file.-->
-<PluginSuite
-    name="PVGeo General Readers and Sources"
-    category="PVGeo General Sources">
-</PluginSuite>
